chatbot-backend/requirements.txt
# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0

# Database
sqlalchemy>=1.4.0
pymongo>=3.12.0
psycopg2-binary>=2.9.5

# Environment and HTTP
python-dotenv>=0.19.0
aiohttp>=3.8.0
requests>=2.28.0

# LangChain and LangGraph
langchain>=0.0.267
langchain-community>=0.0.10
langchain-openai>=0.0.2
langgraph>=0.0.15
openai>=0.27.0

# Vector Search
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# NLP and ML
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.0.0
nltk>=3.7
spacy>=3.5.0
wordcloud>=1.8.0

# Data Processing
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Utilities
pydantic>=1.10.0
tqdm>=4.64.0
